New Feature/consolidated_backend/requirements.txt
# =============================================================================
# ML Backend Dependencies
# USD/COP Forecasting Pipeline
# =============================================================================

# Core Scientific Computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.0.0

# Boosting Models
xgboost>=1.5.0
lightgbm>=3.3.0
catboost>=1.0.0

# Model Persistence
joblib>=1.1.0

# Hyperparameter Optimization
optuna>=3.0.0

# MLflow Integration
mlflow>=2.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Object Storage (MinIO/S3)
boto3>=1.26.0
minio>=7.2.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
pyyaml>=6.0
pytz>=2023.3

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
